requirements.txt
fastapi==0.113.0
uvicorn==0.28.0
sqlalchemy>=2.0.0,<3.0.0
asyncpg==0.29.0
redis>=5.0.0
sentence-transformers==3.0.1
open-clip-torch==2.24.0
torch==2.4.0 --index-url https://download.pytorch.org/whl/cpu
pillow==10.4.0
python-multipart==0.0.6
pydantic-settings==2.1.0
pytest-asyncio==0.23.2
httpx==0.27.0
gunicorn==21.2.0
python-dotenv==1.0.1
sqlmodel==0.0.22
aiosqlite==0.20.0
